pyproject.toml
[tool.poetry]
name = "remindme"
version = "0.1.0"
description = ""
authors = ["Chansoo Song"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.2.0"
flake8 = "^5.0.4"
pre-commit = "^2.20.0"
mkdocs = "^1.4.2"
mkdocs-material = ">=8.0.0,<9.0.0"
mkdocs-simple-blog = "^0.0.5"
mkdocstrings = "^0.20.0"
mkdocs-jupyter = "^0.22.0"
pymdown-extensions = "^9.9.2"
twilio = "^8.0.0"
google-api-python-client = "^2.85.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.0.0"
ipykernel = "^6.22.0"
openai = "^0.27.4"
pyyaml = "^6.0"
flask = "^2.2.3"
django = "^4.2"
django-extensions = "^3.2.1"
ipython = "8.12"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"